Flask_Blog/requirements.txt
aniso8601==9.0.1
apispec==5.2.2
async-timeout==4.0.2
cachelib==0.9.0
click==8.1.3
colorama==0.4.5
configobj==5.0.6
Deprecated==1.2.13
Flask==2.2.2
Flask-Caching==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
flask-smorest==0.39.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
webargs==8.2.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
